langsmith/pyproject.toml
[project]
name = "learn-langsmith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "pydantic>=2.11.2",
]
